requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bestconfig==1.3.5
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
python-magic==0.4.24
PyYAML==6.0
topggpy==1.3.0
typing-extensions==3.10.0.2
Wand==0.6.7
yarl==1.7.2
